# openapi.yaml
openapi: 3.0.3
info:
  title: Lombok Explorer API
  description: |
    API Contract untuk Lombok Explorer — Platform eksplorasi wisata, rekomendasi cerdas, 
    dan perencanaan perjalanan (itinerary planner) pulau Lombok, Nusa Tenggara Barat.
  version: 1.0.0
  contact:
    name: Lombok Explorer Team
    email: support@lombokexplorer.com
servers:
  - url: https://api.lombokexplorer.com/api/v1
    description: Production Server (v1)
  - url: http://localhost:8080/api/v1
    description: Local Development Server (v1)
  - url: http://localhost:8080/v1
    description: Legacy Compatibility Alias Server
paths:
...
  /journals:
    get:
      summary: Get user travel journals
      operationId: getJournals
      tags:
        - Journals
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of travel journals
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/TravelJournalDto'
    post:
      summary: Create new travel journal entry
      operationId: createJournal
      tags:
        - Journals
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateJournalRequest'
      responses:
        '201':
          description: Journal entry created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/TravelJournalDto'

  # ==========================================
  # CHECKLISTS
  # ==========================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Enter JWT Bearer token for authorized requests
...
  schemas:
...
    TravelJournalDto:
      type: object
      required:
        - id
        - title
        - content
        - date
      properties:
        id:
          type: string
          example: "jrn_001"
        title:
          type: string
          example: "Hari Pertama: Sunset Memukau di Bukit Merese"
        content:
          type: string
          example: "Tiba di Lombok siang hari, langsung check-in hotel dan menyewa motor menuju Bukit Merese..."
        locationName:
          type: string
          example: "Bukit Merese, Lombok Selatan"
        photos:
          type: array
          items:
            type: string
          example: ["https://images.unsplash.com/photo-1544644181-1484b3fdfc62?q=80&w=1000"]
        date:
          type: string
          format: date-time
          example: "2026-08-23T08:00:00.000Z"
    CreateJournalRequest:
      type: object
      required:
        - title
        - content
      properties:
        title:
          type: string
          example: "Hari Pertama: Sunset Memukau di Bukit Merese"
        content:
          type: string
          example: "Tiba di Lombok siang hari, langsung check-in hotel..."
        locationName:
          type: string
          example: "Bukit Merese, Lombok Selatan"
        photos:
          type: array
          items:
            type: string
...